prevent ns deletion during migration (#1184)

+++ libsql-server/src/namespace/meta_store.rs
@@ -432,6 +432,11 @@ impl MetaStore {
             let config = sender.borrow().clone();
             let tx = guard.conn.transaction()?;
             if let Some(ref shared_schema) = config.config.shared_schema_name {
+                if crate::schema::db::has_pending_migration_jobs(&tx, shared_schema)? {
+                    return Err(crate::Error::PendingMigrationOnSchema(
+                        shared_schema.clone(),
+                    ));
+                }
                 tx.execute(
                     "DELETE FROM shared_schema_links WHERE shared_schema_name = ? AND namespace = ?",
                     (shared_schema.as_str(), namespace.as_str()),
